Refactored cookie functions to actions.ts.
Testing if this refactor fixes prod error.
